feat: fourceInstantiateWasm
