+++ deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "test": "deno test --doc --parallel --shuffle --allow-net --allow-read=.",
+    "test": "deno test --doc --parallel --shuffle --allow-net=deno.land --allow-read=.",
     "check": "deno check ./mod.ts"
   },
   "importMap": "./import-map.json"
